requirements.txt
# Core Framework
crewai>=0.63.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.4.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.2.0
pdf2image>=1.16.3

# Streamlit UI
streamlit>=1.30.0

# Export
openpyxl>=3.1.2
python-docx>=1.1.0

# HTTP Client
httpx>=0.26.0
aiofiles>=23.2.1
python-multipart>=0.0.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
uuid6>=2024.1.12
